Redundant JOIN
No reason to do a JOIN on a table that isn't being called upon, since the index value is the same in the ovd table.  The ov table just sits there looking like the loser that wasn't invited to the party, but showed up anyway.

Saving a few milliseconds here and there on a query make a big difference when a site gets a lot of traffic.