Getting rid of references to HTTP_SERVER and HTTPS_SERVER
Not needed - this is governed by Apache.

One, stores should NOT run on "http"

TWO, hardcoding in "https" also doesn't make sense in most places, because if Apache isn't configured to use HTTPS, it just breaks things.

I DID hardcode in "https://" for callback URL's, though, because those services won't talk to us otherwise, but beyond that, it's really the role of the server admin to set this up correctly.