Add enable parameters in DB\Query, and added optional logging
This was only meant to add Query logging, but saw that the top level DB adapter ignores parameters/doesn't pass them to the PDO adapter.

Now it does, so it should be easier to enable prepared statments and get rid of all of the $this->db->escape, etc.

Also, added query logger. Only runs if DEBUG_SQL is defined and set to 1