Modify PHP error logging
Current implementation leaves it to the admin to turn on logging in the UI.

Now there are 3 constants in config.php that turn logging on and off.

PHP_ERROR_DISPLAY (shows the error on the screen)

PHP_ERROR_LOG (writes the error to a text file)

PHP_ERROR_HALT (stops executing the script).

The new feature is to halt execution, to stop letting errors slip silently by.