Fix */controller/startup/error.php
both of thse were looking to the $this->config->get for the error log filename

Have now hardcoded it as DIR_LOGS . 'error.log'

We COULD move the name into config.php or elsewhere, but I'm fine not letting site admin change it, only webmaster, since they're the only one that should be able to do that.

Also, both of these controllers are identical - there should be a 'common' directory to load common controllers out of and not need duplication.