Refactor
Refactored catalog / featured products extension

/public/catalog/controller/extension/module/featured.php

Now makes a single query for ALL of the featured products, rather than one query per featured product. Not a big time saver, but it'll be something that gets repeated in other modules/features of the site. This was just an easy place to start.

(NOTE: I made the database call a part of the controller file, which I'm treating as a plug-in, rather than make a separate model, or make changes to the the products controller. If other moduels are using the same select, THEN i'll move to the main model, but otherwise, I'm happy with this setup)