Delete remnants of OC Database Abstraction Layer
This will work just as easily as Opencarts old method - would be able to support any database that PDO supports, except for SQL specific issues, which is exactly the same as Opencart is currently.

Silly holdover that MySQLi still is used once PDO came in. And equally silly that it pretentds to support MSSQL and POSTGRES.

EIther way, if it would have run under other databases in the past, it still will, with slight code modifications